Handle ANR due to blocking semaphores
If DLS was invoked twice using Intents (renewRights or
Browser->WebInitiatorClick) and the phone also is out of service
at the same time, then DLS would ANR because HttpThread is using
latch.wait for a long time while a new item is added to database
(and thus requiring a lock). These locks will wait for each other
and thus cause the ANR. The fix is to do the second lock request
in a new Thread (thus not blocking the main service thread).